refactor: move getStargateQuote from api.ts to stargate.ts and ABI to abis/
Since getStargateQuote is now a direct contract call (not an API call),
moved it to lib/utils/stargate.ts alongside the other Stargate utilities
and extracted the quoteSend ABI to lib/abis/StargateOFT.ts. Updated all
6 consumer imports accordingly.
