+++ hooks/useBridgeToCard.ts
@@ -11,7 +11,6 @@ import { useActivityActions } from '@/hooks/useActivityActions';
 import BridgePayamster_ABI from '@/lib/abis/BridgePayamster';
 import { CardDepositManager_ABI } from '@/lib/abis/CardDepositManager';
 import { track } from '@/lib/analytics';
-import { getStargateQuote } from '@/lib/api';
 import {
   ADDRESSES,
   EXPO_PUBLIC_CARD_FUNDING_CHAIN_ID,
@@ -20,7 +19,7 @@ import {
 import { executeTransactions, USER_CANCELLED_TRANSACTION } from '@/lib/execute';
 import { StargateQuoteParams, Status, TransactionType } from '@/lib/types';
 import { getCardDepositTokenAddress, getCardFundingAddress } from '@/lib/utils';
-import { getStargateChainId } from '@/lib/utils/stargate';
+import { getStargateChainId, getStargateQuote } from '@/lib/utils/stargate';
 
 import { useCardContracts } from './useCardContracts';
 import { useCardDetails } from './useCardDetails';
